+++ pytket/docs/changelog.rst
@@ -12,8 +12,9 @@ Features:
 
 * Add ``circuit_name`` property to ``CircBox``.
 * Enable pickling of ``Bit`` objects.
-* New ``Transform.PushCliffordsThroughMeasures()`` that reimplements Clifford
-  subcircuits before end of circuit measurements by measurement reduction.
+* New optimisation ``Transform.PushCliffordsThroughMeasures()`` and pass 
+  ``CliffordPushThroughMeasures`` that optimises Clifford subcircuits 
+  before end of circuit measurement gates.
 
 Fixes: